migrate eval cmds to expr module
